requirements.txt
Flask>=2.0.0
PyJWT>=2.0.0
requests>=2.28.0